move filedistributorrpc.def again
* filedistributorrpc.def now lives in fileacquirer module
* is used directly by java build
* is packaged into a library for use in C++
