+++ CMakeLists.txt
@@ -30,6 +30,7 @@ add_subdirectory(documentapi)
 add_subdirectory(fastlib)
 add_subdirectory(fastos)
 add_subdirectory(fbench)
+add_subdirectory(fileacquirer)
 add_subdirectory(filedistribution)
 add_subdirectory(fnet)
 add_subdirectory(frtstream)
