+++ configdefinitions/src/vespa/CMakeLists.txt
@@ -62,5 +62,3 @@ vespa_generate_config(configdefinitions zookeeper-server.def)
 install(FILES zookeeper-server.def DESTINATION var/db/vespa/config_server/serverdb/classes)
 vespa_generate_config(configdefinitions zookeepers.def)
 install(FILES zookeepers.def DESTINATION var/db/vespa/config_server/serverdb/classes)
-vespa_generate_config(configdefinitions filedistributorrpc.def)
-install(FILES filedistributorrpc.def DESTINATION var/db/vespa/config_server/serverdb/classes)
