+++ fileacquirer/src/main/resources/configdefinitions/filedistributorrpc.def
