+++ fileacquirer/src/vespa/fileacquirer/.gitignore
@@ -1,0 +1,2 @@
+config-filedistributorrpc.cpp
+config-filedistributorrpc.h
