+++ fileacquirer/src/vespa/fileacquirer/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_library(fileacquirer STATIC
+    SOURCES
+    DEPENDS
+)
+vespa_generate_config(fileacquirer filedistributorrpc.def)
+install(FILES filedistributorrpc.def DESTINATION var/db/vespa/config_server/serverdb/classes)
