+++ fileacquirer/src/vespa/fileacquirer/filedistributorrpc.def
@@ -1,0 +1,1 @@
+../../main/resources/configdefinitions/filedistributorrpc.def
